revamp stage 5: derivation.py, provenance DAG to prose
The 340 lines of delta abstraction, loop-structure recovery,
anti-unification folding, closed forms and rendering move out of
pair.py into a module whose docstring states the three-pass design
and the no-drift guarantee. pair.py re-imports the historical names,
so tests that reach for _delta_steps and friends are untouched.
